+++ tests/testthat/test-runs.R
@@ -112,3 +112,30 @@ test_that("validates `start`", {
   expect_error(vec_locate_runs(1, start = NA), "single `TRUE` or `FALSE`")
   expect_error(vec_locate_runs(1, start = c(TRUE, TRUE)), "single `TRUE` or `FALSE`")
 })
+
+# vec_detect_runs --------------------------------------------------------------
+
+test_that("can detect run starts", {
+  expect_identical(
+    vec_detect_runs(c(1, 3, 3, 1, 5, 5, 6)),
+    c(TRUE, TRUE, FALSE, TRUE, TRUE, FALSE, TRUE)
+  )
+})
+
+test_that("can detect run ends", {
+  expect_identical(
+    vec_detect_runs(c(1, 3, 3, 1, 5, 5, 6), start = FALSE),
+    c(TRUE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE)
+  )
+})
+
+test_that("works with size zero input", {
+  expect_identical(vec_detect_runs(integer(), start = TRUE), logical())
+  expect_identical(vec_detect_runs(integer(), start = FALSE), logical())
+})
+
+test_that("validates `start`", {
+  expect_error(vec_detect_runs(1, start = "x"), "single `TRUE` or `FALSE`")
+  expect_error(vec_detect_runs(1, start = NA), "single `TRUE` or `FALSE`")
+  expect_error(vec_detect_runs(1, start = c(TRUE, TRUE)), "single `TRUE` or `FALSE`")
+})